started fish tally, updated readme
