+++ README.md
@@ -2,35 +2,8 @@
 ## Introduction
 Each year between July and September brown bears in the Katmai National Park congregate along the Brooks River.  This river is one of the main rivers that Sockeye Salmon swim up to spawn.  The brown bears wait in the shallows of the river and catch the salmon as they jump over rocks.  Want to see it in [ACTION](https://explore.org/livecams/brown-bears/brown-bear-salmon-cam-brooks-falls)?
 
-You have been tasked with creating a site scientists and park ranger can use to track these bears.
+This is a site scientists and park ranger can use to track these bears.
 
-## Get Started
-
-1. Clone this repo
-```
-$ git clone github-url
-$ cd repo-name
-```
-2. Follow the instructions from the ["Starting the Project" Instructions](https://github.com/nss-nightclass-projects/webpack-template/blob/master/README.md#starting-the-project) to ensure you make all the updates to get the project started.
-
-3. Install Packages
-
-```
-$ sudo npm install
-$ npm start
-```
-
-## Planning Requirements
-* Create a github project on this repo
-* Create feature tickets for this assignment
-* Create wireframes for this project and attach them to the appropriate project ticket.
-
-## Technical Requirements
-* You must use [Boostrap](https://getbootstrap.com/) to style your page components.
-* Your JS file should be comprised of functions, no actions should happen in your code outside of a function
-* Your HTML and JS should all have proper indentations
-* You should be have Modular Code
-* Your JS code should be eslint error and warning free
 
 ## User stories
 * As a user, when the page loads, I should see a bootstrap form that allows me to enter in a bear name and image url for a new bear.
@@ -52,17 +25,13 @@ $ npm start
 Each year in early september one bear is awarded the Fattest Bear award.  This award goes to the bear who gains the most weight between July and September.  Create an awards section on your application and award GOLD to the bear who has eaten the most fish, SILVER to the bear with the second most eaten fish, and BRONZE to the bear with the third most eaten fish.  Where applicable your application should show two winners for an award if the number of fish is tied.
 
 
-NOTE: you will have to update this to meet your needs. DO NOT copy and paste this template without making updates to it.
-
-# This is the name of my Project  [![Netlify Status](https://api.netlify.com/api/v1/badges/4ab7e730-7ed3-4cfd-a988-66195e79a991/deploy-status)](https://app.netlify.com/sites/drt-sortinghat/deploys)
-<!-- update the netlify badge above with your own badge that you can find at netlify under settings/general#status-badges -->
+# Bear Trackers Unite!  [![Netlify Status](https://api.netlify.com/api/v1/badges/ab1d40f0-ae98-4bb6-aa84-9e803362dfd9/deploy-status)](https://app.netlify.com/sites/dc-bear-watcher/deploys)
 
 This is an app mainly used by rangers and scientist to track and research bears in the Katmai National Park. Each year between July and September brown bears congregate along the Brooks River. This river is one of the main rivers that Sockeye Salmon swim up to spawn. The brown bears wait in the shallows of the river and catch the salmon as they jump over rocks. 
 
-[View App](#your-link)
+[View App](dc-bear-watcher.netlify.app)
 
 ## Get Started
-PLACE CODE SNIPPET HERE
 
 ## About the User
 - The ideal user for this application is a teacher, park ranger, scientist or nature enthueist
@@ -75,20 +44,17 @@ PLACE CODE SNIPPET HERE
 * As a user, when I click the submit button on the new bear form a bear should be added to the page.
 * As a user, when there are bears that are being tracked (ie after I have added one), I should see all their information displayed on a card.
 
-## Video Walkthrough of APP NAME <!-- A loom link is sufficient -->
+## Loom video Walkthrough of Bear Watchers Unite!
 
 
 ## Relevant Links <!-- Link to all the things that are required outside of the ones that have their own section -->
-- [Check out the deployed site](#your-link)
-- [Wireframes](#your-link)
-- [Project Board](#your-link)
+- [Check out the deployed site](dc-bear-watcher.netlify.app)
+- [Figma WireFrame](https://www.figma.com/file/6VTWlVFWrUrVviW2rD6iXo/Bear-Watcher?node-id=0%3A1)
+- [Project Board](https://github.com/danicrosby/bear-watcher/projects)
 
-## Code Snippet <!-- OPTIONAL, but doesn't hurt -->
-PLACE CODE SNIPPET HERE
+## Project Screenshots
+<img width="1148" alt="Bear Watcher App Overview" src="./images/Overview.png">
 
-## Project Screenshots <!-- These can be inside of your project. Look at the repos from class and see how the images are included in the readme -->
-<img width="1148" alt="Your Alt" src="your-link.png">
 
 ## Contributors
 - Dani Crosby: https://github.com/danicrosby
-
