+++ src/javascripts/componenets/bearData.js
@@ -1,23 +1,38 @@
 const bears = [
   {
+    id: 0,
     name: 'Fozzy',
-    imgUrl: 'http://3.bp.blogspot.com/_Dp2Dd4Y_YuE/SJkMdNEXcUI/AAAAAAAAARg/jfUq7PXiiS0/s400/Brown+bear+piscivory+blog+copy.jpg',
+    imgUrl: 'http://3.bp.blogspot.com/_Dp2Dd4Y_YuE/SJkMdNEXcUI/AAAAAAAAARg/jfUq7PXiiS0/s400/Brown+bear+piscivory+blog',
+    caught: 2,
+    missed: 5,
   },
   {
+    id: 1,
     name: 'Teddy',
     imgUrl: 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTanjD8u20zCc7UBGr62bDgqTvVuNt_uwotLg&usqp=CAU',
+    caught: 2,
+    missed: 5,
   },
   {
+    id: 2,
     name: 'Smokey',
     imgUrl: 'https://i.pinimg.com/originals/04/c1/fd/04c1fd5d24366794970a49edbc57eeb0.jpg',
+    caught: 2,
+    missed: 5,
   },
   {
+    id: 3,
     name: 'Fuzzy Wuzzy',
     imgUrl: 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQkxekTymMIHA0cFD1TBY9mMbxzEnHdQwWF0w&usqp=CAU',
+    caught: 2,
+    missed: 5,
   },
   {
+    id: 4,
     name: 'Terrance',
     imgUrl: 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcSuFnV11UZ1V0TF26NKUiBVczkPfYAldX8AMA&usqp=CAU',
+    caught: 2,
+    missed: 5,
   },
 ];
